requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
boto3>=1.34.0
arxiv==2.1.0
PyMuPDF==1.23.21
pytest==8.0.2
pytest-mock==3.12.0
python-dotenv==1.0.1